Make a unit test pass both with and without optimization
Made the smallest (for now, rather than the cleanest) change to make
the failing test pass without optimization. Motivation: to emphasize
the mistake.

The base-10 logarithm L of the maximal value is calculated for exactly
one purpose: 1+floor(L) is a proxy for the number of digits to the left
of the decimal point in that value's decimal representation. Although
this logarithm is a floating-point quantity, only its integer part is
actually wanted. Therefore, it should be immediately converted to an
integer. To perform inexact floating-point operations upon it before
conversion to integer is to invite preventible inexactness into the
calculation.

In this context, truncation is appropriate: no choice need be made
between floor() and trunc() semantics, because only a natural number is
wanted. Naturally, a later revision of the changed line:
    double d = static_cast<int>(std::log10(widest));
will store the integer value in an integer variable.

+++ miscellany.cpp
@@ -135,7 +135,7 @@ int scale_power(int max_power, double min_value, double max_value)
         return 0;
         }
 
-    double d = std::log10(widest);
+    double d = static_cast<int>(std::log10(widest));
     d = std::floor(d / 3.0);
     int k = 3 * static_cast<int>(d);
     k = k - 6;
